api/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12.4,<3.13
djangorestframework-simplejwt>=5.1.0, <5.2.0
psycopg2>=2.8